% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_idle_simulation.R
\name{analyze_idle_simulation}
\alias{analyze_idle_simulation}
\title{Analyze Idle Mechanic Simulation}
\usage{
analyze_idle_simulation(results)
}
\arguments{
\item{results}{The results from the simulation.}
}
\value{
A data frame with the final statistics and interpretations.
}
\description{
This function analyzes the results of the idle mechanic simulation.
}
