% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_event.R
\name{apply_event}
\alias{apply_event}
\title{Apply Event}
\usage{
apply_event(user_data, event)
}
\arguments{
\item{user_data}{A tibble containing user data.}

\item{event}{A tibble representing the event.}
}
\value{
A tibble with updated user data.
}
\description{
This function applies an event to the user data.
}
