% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_resource_generation.R
\name{apply_resource_generation}
\alias{apply_resource_generation}
\title{Apply Resource Generation}
\usage{
apply_resource_generation(user_data, generator, time_period)
}
\arguments{
\item{user_data}{A tibble containing user data.}

\item{generator}{A tibble representing the resource generator.}

\item{time_period}{The time period over which resources are generated.}
}
\value{
A tibble with updated user data.
}
\description{
This function applies resource generation to user data.
}
