% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_time_to_complete_tech_tree.R
\name{average_time_to_complete_tech_tree}
\alias{average_time_to_complete_tech_tree}
\title{Calculate Average Time to Complete Tech Tree}
\usage{
average_time_to_complete_tech_tree(user_data)
}
\arguments{
\item{user_data}{A tibble containing user data with a time component.}
}
\value{
The average time to complete the tech tree.
}
\description{
This function calculates the average time taken by users to complete the tech tree.
}
