% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balancing_functions.R
\docType{data}
\name{balancing_functions}
\alias{balancing_functions}
\title{Balancing Functions}
\format{
An object of class \code{list} of length 4.
}
\usage{
balancing_functions
}
\arguments{
\item{x}{The input value.}

\item{...}{Additional parameters for the balancing function.}
}
\value{
The output value after applying the balancing function.
}
\description{
A collection of balancing functions for the idle mechanic simulation.
}
\keyword{datasets}
