% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_quest.R
\name{complete_quest}
\alias{complete_quest}
\title{Complete Quest}
\usage{
complete_quest(user_data, quest)
}
\arguments{
\item{user_data}{A tibble containing user data.}

\item{quest}{A tibble representing the quest.}
}
\value{
A tibble with updated user data and quest completion status.
}
\description{
This function completes a quest if the requirement is met.
}
