% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_action.R
\name{define_action}
\alias{define_action}
\title{Define Action}
\usage{
define_action(
  action_name,
  cost_currency1 = 0,
  cost_currency2 = 0,
  perceived_value = 1,
  public_order_change = 0,
  happiness_change = 0,
  xp_reward = 0,
  level_requirement = 1
)
}
\arguments{
\item{action_name}{The name of the action.}

\item{cost_currency1}{Cost in currency 1.}

\item{cost_currency2}{Cost in currency 2.}

\item{perceived_value}{Perceived value of the action.}

\item{public_order_change}{Change in public order.}

\item{happiness_change}{Change in happiness.}

\item{xp_reward}{XP reward for completing the action.}

\item{level_requirement}{Level requirement to perform the action.}
}
\value{
A tibble representing the action.
}
\description{
This function defines an action with its costs, rewards, and requirements.
}
