% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_battle.R
\name{define_battle}
\alias{define_battle}
\title{Define Battle}
\usage{
define_battle(
  battle_name,
  required_level,
  success_rate,
  reward_currency1 = 0,
  reward_currency2 = 0,
  xp_reward = 0
)
}
\arguments{
\item{battle_name}{The name of the battle.}

\item{required_level}{Minimum player level required to participate.}

\item{success_rate}{Probability of winning the battle.}

\item{reward_currency1}{Reward in currency 1 for winning.}

\item{reward_currency2}{Reward in currency 2 for winning.}

\item{xp_reward}{XP reward for winning.}
}
\value{
A tibble representing the battle.
}
\description{
This function defines a battle scenario.
}
