% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_building.R
\name{define_building}
\alias{define_building}
\title{Define Building}
\usage{
define_building(
  building_name,
  construction_cost_currency1,
  construction_cost_currency2,
  tech_tree
)
}
\arguments{
\item{building_name}{The name of the building.}

\item{construction_cost_currency1}{Construction cost in currency 1.}

\item{construction_cost_currency2}{Construction cost in currency 2.}

\item{tech_tree}{The tech tree associated with the building.}
}
\value{
A tibble representing the building.
}
\description{
This function defines a building with its construction costs and associated tech tree.
}
