% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_event.R
\name{define_event}
\alias{define_event}
\title{Define Event}
\usage{
define_event(
  event_name,
  currency1_change = 0,
  currency2_change = 0,
  public_order_change = 0,
  happiness_change = 0,
  xp_change = 0
)
}
\arguments{
\item{event_name}{The name of the event.}

\item{currency1_change}{Change in currency 1.}

\item{currency2_change}{Change in currency 2.}

\item{public_order_change}{Change in public order.}

\item{happiness_change}{Change in happiness.}

\item{xp_change}{Change in XP.}
}
\value{
A tibble representing the event.
}
\description{
This function defines an event with its effects.
}
