% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_quest.R
\name{define_quest}
\alias{define_quest}
\title{Define Quest}
\usage{
define_quest(
  quest_name,
  requirement,
  reward_currency1 = 0,
  reward_currency2 = 0,
  reward_xp = 0
)
}
\arguments{
\item{quest_name}{The name of the quest.}

\item{requirement}{The requirement to complete the quest.}

\item{reward_currency1}{Reward in currency 1.}

\item{reward_currency2}{Reward in currency 2.}

\item{reward_xp}{Reward in XP.}
}
\value{
A tibble representing the quest.
}
\description{
This function defines a quest with its requirements and rewards.
}
