% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_resource_generator.R
\name{define_resource_generator}
\alias{define_resource_generator}
\title{Define Resource Generator}
\usage{
define_resource_generator(
  generator_name,
  resource_rate1 = 0,
  resource_rate2 = 0
)
}
\arguments{
\item{generator_name}{The name of the generator.}

\item{resource_rate1}{Rate of resource 1 generation.}

\item{resource_rate2}{Rate of resource 2 generation.}
}
\value{
A tibble representing the resource generator.
}
\description{
This function defines a resource generator.
}
