% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_tech_tree.R
\name{define_tech_tree}
\alias{define_tech_tree}
\title{Define Tech Tree}
\usage{
define_tech_tree(...)
}
\arguments{
\item{...}{Tech nodes to be included in the tech tree.}
}
\value{
A tibble representing the tech tree.
}
\description{
This function defines a tech tree composed of multiple tech nodes.
}
