% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_inventory.R
\name{initialize_inventory}
\alias{initialize_inventory}
\title{Initialize Inventory}
\usage{
initialize_inventory(items = list())
}
\arguments{
\item{items}{A named list of items with their initial quantities.}
}
\value{
A tibble representing the inventory.
}
\description{
This function initializes an inventory for a user.
}
