% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_user.R
\name{initialize_user}
\alias{initialize_user}
\title{Initialize a User}
\usage{
initialize_user(
  user_id = 1,
  initial_currency1 = 100,
  initial_currency2 = 50,
  initial_public_order = 50,
  initial_happiness = 50,
  initial_xp = 0,
  initial_level = 1,
  ...
)
}
\arguments{
\item{user_id}{User ID.}

\item{initial_currency1}{Initial amount of currency 1.}

\item{initial_currency2}{Initial amount of currency 2.}

\item{initial_public_order}{Initial public order value.}

\item{initial_happiness}{Initial happiness value.}

\item{initial_xp}{Initial XP.}

\item{initial_level}{Initial level.}

\item{...}{Additional attributes.}
}
\value{
A tibble representing the user data.
}
\description{
This function initializes a user with default or specified attributes.
}
