% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_inventory.R
\name{modify_inventory}
\alias{modify_inventory}
\title{Modify Inventory}
\usage{
modify_inventory(inventory, item_name, quantity_change)
}
\arguments{
\item{inventory}{A tibble representing the inventory.}

\item{item_name}{The name of the item to modify.}

\item{quantity_change}{The change in quantity of the item.}
}
\value{
A tibble with updated inventory.
}
\description{
This function modifies the inventory by adding or removing items.
}
