% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scenarios.R
\name{run_scenarios}
\alias{run_scenarios}
\title{Run Scenarios}
\usage{
run_scenarios(user, building, tech_tree, scenarios)
}
\arguments{
\item{user}{A tibble containing initial user data.}

\item{building}{A tibble representing the building.}

\item{tech_tree}{A tibble representing the tech tree.}

\item{scenarios}{A list of scenario parameters.}
}
\value{
A list containing the results of each scenario.
}
\description{
This function runs multiple scenarios of simulations.
}
