% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_action}
\alias{simulate_action}
\title{Simulate Action}
\usage{
simulate_action(user_data, action)
}
\arguments{
\item{user_data}{A tibble containing user data.}

\item{action}{A tibble representing the action.}
}
\value{
A tibble with updated user data.
}
\description{
This function simulates an action and updates user data.
}
