% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_battle.R
\name{simulate_battle}
\alias{simulate_battle}
\title{Simulate Battle}
\usage{
simulate_battle(user_data, battle)
}
\arguments{
\item{user_data}{A tibble containing user data.}

\item{battle}{A tibble representing the battle.}
}
\value{
A tibble with updated user data.
}
\description{
This function simulates a battle and updates user data based on the outcome.
}
