% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_building_construction.R
\name{simulate_building_construction}
\alias{simulate_building_construction}
\title{Simulate Building Construction}
\usage{
simulate_building_construction(user_data, building)
}
\arguments{
\item{user_data}{A tibble containing user data.}

\item{building}{A tibble representing the building.}
}
\value{
A tibble with updated user data.
}
\description{
This function simulates the construction of a building and updates user data.
}
