% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_idle_mechanic.R
\name{simulate_idle_mechanic}
\alias{simulate_idle_mechanic}
\title{Simulate Idle Mechanic}
\usage{
simulate_idle_mechanic(
  balancing_func,
  goal_increase_func,
  simulation_time,
  initial_money_per_sec,
  initial_goal,
  initial_reset_boost,
  initial_boost_currency_per_reset,
  reset_boost_growth_rate,
  boost_currency_growth_rate,
  func_params = list()
)
}
\arguments{
\item{balancing_func}{The balancing function to use.}

\item{goal_increase_func}{The function to increase the goal after each reset.}

\item{simulation_time}{The total simulation time in seconds.}

\item{initial_money_per_sec}{The initial money generation rate per second.}

\item{initial_goal}{The initial monetary goal for market reset.}

\item{initial_reset_boost}{The initial permanent production boost per reset.}

\item{initial_boost_currency_per_reset}{The initial boost currency earned per market reset.}

\item{reset_boost_growth_rate}{The growth rate of reset boost per reset.}

\item{boost_currency_growth_rate}{The growth rate of boost currency per reset.}

\item{func_params}{A list of parameters for the balancing function.}
}
\value{
A list containing time series data and final statistics.
}
\description{
This function simulates an idle mechanic with given balancing functions and parameters.
}
