% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_tech_progression}
\alias{simulate_tech_progression}
\title{Simulate Tech Progression}
\usage{
simulate_tech_progression(user_data, tech_tree)
}
\arguments{
\item{user_data}{A tibble containing user data.}

\item{tech_tree}{A tibble representing the tech tree.}
}
\value{
A tibble with updated user data.
}
\description{
This function simulates the progression through a tech tree and updates user data.
}
