% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{update_user_level}
\alias{update_user_level}
\title{Update User Level}
\usage{
update_user_level(user_data)
}
\arguments{
\item{user_data}{A tibble containing user data.}
}
\value{
A tibble with updated user level.
}
\description{
This function updates the user's level based on their XP.
}
