% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_currency.R
\name{visualize_currency}
\alias{visualize_currency}
\title{Visualize Currency Changes Over Time}
\usage{
visualize_currency(user_data)
}
\arguments{
\item{user_data}{A tibble containing user data with a time component.}
}
\value{
A ggplot object showing the currency changes over time.
}
\description{
This function plots the changes in currency over time.
}
