% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_idle_simulation.R
\name{visualize_idle_simulation}
\alias{visualize_idle_simulation}
\title{Visualize Idle Mechanic Simulation}
\usage{
visualize_idle_simulation(time_series)
}
\arguments{
\item{time_series}{The time series data from the simulation.}
}
\value{
A ggplot object showing the simulation results.
}
\description{
This function visualizes the results of the idle mechanic simulation.
}
