% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_tech_tree.R
\name{visualize_tech_tree}
\alias{visualize_tech_tree}
\title{Visualize Tech Tree Progression}
\usage{
visualize_tech_tree(user_data, tech_tree)
}
\arguments{
\item{user_data}{A tibble containing user data with tech tree progression.}

\item{tech_tree}{The tech tree structure.}
}
\value{
A ggplot object showing the tech tree progression.
}
\description{
This function plots the progression through the tech tree.
}
